+++ src/sweet_tooth/endpoint/module/middleware.clj
@@ -2,6 +2,7 @@
   (:require [duct.core :as duct]
             [duct.middleware.buddy :as dbuddy]
             [integrant.core :as ig]
+            [sweet-tooth.endpoint.handler :as eh]
             [sweet-tooth.endpoint.middleware :as em]))
 
 (def ^:private middleware-config-base
@@ -11,6 +12,7 @@
    ::em/format-response    {}
    ::em/format-exception   {}
    ::em/stacktrace-log     {}
+   ::em/not-found          {:error-handler (ig/ref ::eh/index.html)}
    ::dbuddy/authentication ^:displace {:backend :session}})
 
 (def ^:private middleware-config-dev
@@ -21,7 +23,8 @@
 
 (defmethod ig/init-key :sweet-tooth.endpoint.module/middleware [_ {:keys [exclude]}]
   (fn [{:keys [:duct.core/environment] :as config}]
-    (let [selected-middlewares (remove (set exclude) [::dbuddy/authentication
+    (let [selected-middlewares (remove (set exclude) [::em/not-found
+                                                      ::dbuddy/authentication
                                                       ::em/format-response
                                                       ::em/merge-params
                                                       ::em/stacktrace-log
@@ -36,8 +39,7 @@
                                           (comp vec (partial remove #{(ig/ref :duct.middleware.web/stacktrace)})))]
       (duct/merge-configs
        (select-keys middleware-config selected-middlewares)
-       {:duct.middleware.web/not-found           {:error-handler (ig/ref :sweet-tooth.endpoint.handler/index.html)}
-        :sweet-tooth.endpoint.handler/index.html {}}
+       {::eh/index.html {}}
        config
        {:duct.handler/root {:middleware (with-meta (mapv ig/ref prepend-middlewares) {:prepend true})}}
        {:duct.handler/root {:middleware (mapv ig/ref append-middlewares)}}))))